% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypeR.R
\name{hypeROverrepEnrichFunction}
\alias{hypeROverrepEnrichFunction}
\title{hypeROverrepEnrichFunction()}
\usage{
hypeROverrepEnrichFunction(
  sigSymbol,
  gsetNames = NULL,
  gsetList = NULL,
  gsetListCategory = "customize",
  gsetListSubCategory = "customize",
  test = NULL
)
}
\arguments{
\item{sigSymbol}{A character contains signature symbols.}

\item{gsetNames}{A dataframe or matrix to specify which geneset(s) to use. required columns including "species", "category", "subcategory". See hypeR::msigdb_info() to see available genesets.}

\item{gsetList}{A list of genesets. the name of each entry should be the geneset name. e.g. valid output of readGsetList() function.}

\item{gsetListCategory}{A single character string; need this to rbind result from msigDB gset and customize gset.}

\item{gsetListSubCategory}{A single character string; need this to rbind result from msigDB gset and customize gset.}

\item{test}{"overrep" or "enrich" for over-representation analysis and rank-based enrichment analysis}
}
\value{
A dataframe of over-representation analysis result.
}
\description{
Perform over-representative analysis using hypeR package
hypeROverrepEnrichFunction() allow input BOTH msigdb gsets (a df with names) as gsetNames
and a customized gset list (a list) as gsetList,
and output result in a df.
last updated 04/21/2020 by vmli
}
\examples{

testSigSymbol <- c("TNMD", "FUCA2", "C1orf112", "FGR", "MAD1L1", "NFYA",
"STPG1", "CFTR", "LAS1L", "ENPP4", "SEMA3F", "ANKIB1", "KRIT1", "RAD52",
"MYH16", "BAD", "LAP3", "CD99", "HS3ST1", "AOC1", "WNT16", "SNX11",
"TMEM176A", "M6PR")
gsetNames <- data.frame(species = "Homo sapiens",
category = c("C2", "C4"),
subcategory = c("CP:KEGG", "CGN"))
hypeROverrepEnrichFunction(testSigSymbol, gsetNames, test = "overrep")


}
