% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypeR.R
\name{readGsetList}
\alias{readGsetList}
\title{readGsetList()}
\usage{
readGsetList(file = NULL)
}
\arguments{
\item{file}{A txt file with geneset inside. each line should be a gene set.
first value be the gene set name.
genes should be tab seperated.}
}
\value{
A list
}
\description{
read geneset from a txt file,
e.g. a file downloaded from Enrichr https://amp.pharm.mssm.edu/Enrichr/#stats
last updated 04/21/2020 by vmli
}
