% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFun.R
\name{testFun}
\alias{testFun}
\title{a test function to see how it works. Meows}
\usage{
testFun(num = NULL, text = NULL, cat = TRUE)
}
\arguments{
\item{num}{A number}

\item{text}{A character string}

\item{cat}{either T or F}
}
\value{
A list
}
\description{
type a number, and it will gives you its square and cubic value.
Or a text, it will gives you the length of it.
}
\examples{
testFun(num=3, cat=TRUE)
testFun(5, "oh hello", cat=TRUE)
}
